refactor(pool): move the pipe-error grace handling into the forks worker
